Deleted the Structure folder to remove ability to minify doors
The Structures.xml file only held the patches for doors. I've seen users complain about issues with doors involving raiders stealing their doors which is obviously a problem. And I personally feel doors are a trivial thing to be minifiable in the first place. I don't currently have plans to implement patches for other structures either and so deleted the entire file until the need arises.