Bugfix: duplicate filepath overwrite and other small tweaks
The compatibility update for ImprovedWastepackAtomizer was made using the same filename as the base mod "Misc.xml". This was overwriting the Misc.xml file in the base mod Common/Misc.xml, thus losing the changes to the original misc buildings.

Also corrected the thing categories for the wastepack atomizer variants.

Removed the "MayRequire" attribures as the LoadFolders.xml makes it redundant.